Continued merging components from one code base to another
I continued moving the components that could be reused from the previous code base to this one. I as well fixed some styling and made some adjustments. I fully set up the Tailwind CSS. Moved all of the Tailwind variables from the previous repository to this one. I as well proceeded on with setting types and interfaces in order to fully embrace type script