The alert box and form validation
I did the alert dialog box, that communicates with the user when "next" is pressed on the contacts info page. It tells the user that this is the end of the order and by pressing yes, they will send in the query and finish the order. I created a component that holds the dialog alert box. The component takes two properties. the setOpenDialog and openDialog. They are beeing passed from the parent component and controll the toggle up and down of the dialog.
I as well started working on the contact-info form validation.  It needs to display the missing field with red. I have a problem with makling it work. I am passing an object to each form item, the object contains "value" and "hasError". But for some reason, when I trigger the function that dispatches the user input, it overwrites the whole object. Need to figure out how to solve it.