The form validation
Made the form validation of the contacts info page. Now, when a user misses a field, it turns red. to notify the user, that the field is required. I as well added an alert component from MUI, to communicate with the user.
The form validation is being handled by a react useReducer hook, and the logic for the reducer is written in the ContactsInformationContext context component. The user interactions with the form are handled by a massive if statement block in the ContactsPage.tsx page component file. For example, when the user provides all the fields but with not supported format, it throws an error and alert box, saying that the user needs to provide a valid format.
The error alert is being handed by a component, i developed, specifically for the error alert. A state variable, that is managed by the if statement block, was initialized in the ContactsPage file, the variable controls the body text of the alert.