Small fixes and wrapping up
Imported and initialized the alert component in the product selection page. Now, on error it uses the alert component instead of the window.alert(). One of the components is for warnings, the other one for errors. Removed some unnecessary code I saw, as well as added some comments for clarity.