One of the last fixes
Some minor polishings, such as the cursor appearance on certain elements, animation durations and animation delays.
Also re-wrote the logic for the direct user input amount handler. Now the user can type in 0, without the product component disappearing. I did that by assigning a middleman state variable, that takes the input and then updates the object of the product.