The product description po-over
I designed and developed a component, that will be used as a pop-over showing the description of a product. I used Material design for the pop-over UI component, developed the logic and layout of the body of the component myself. The component is an atomic component and the file could be found in Atomic componentss