Re-wrote the logic for delivery date
I re-factored the logic that handles the change of delivery date. Since there was a problem with the default value, there was no default value of delivery date, when the user doesn't select a date. I needed to change that. I decided to handle the issue with the most simple way I could figure out. I used the useEffect hook. once the component is rendered on screen,. the useEffect hook triggers the function that handles change on the calendar. The function is getting the first possible date as a default date (in this case 14 days from the current date).