Fixed the issue with adding products
Wrote a new logic for handling the adding of a product to the basket. Since it was inconvinient for the user to need to press two times on two different locations on the page in order to add the product to the basket, I decided to change the logic. Now, once the user selects a product variation, the product is being added to the basket. I re-wrote the  handleVariationChoice. Instead of using one function for updating the local state, and another function for validation and adding the state of the product to the basket, I combined them into one. Now there is only one function that handles the whole logic. Thus I increase performance and also remove unnecessary code.